third_party/blink/web_tests/platform/generic/fast/css/annotated-regions-expected.txt

This tests that transformed annotated regions have the correct bounds.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS draggableRegions() is ''
PASS nonDraggableRegions() is ''
PASS draggableRegions() is '20,10+100x60'
PASS nonDraggableRegions() is ''
PASS draggableRegions() is '45,25+50x30'
PASS nonDraggableRegions() is ''
PASS draggableRegions() is '45,25+50x30'
PASS nonDraggableRegions() is '45,25+30x10'
PASS draggableRegions() is '20,10+100x60'
PASS nonDraggableRegions() is '20,10+60x20'
PASS successfullyParsed is true

TEST COMPLETE
